+++ apps/prairielearn/src/pages/instructorStudents/instructorStudents.html.tsx
@@ -28,7 +28,7 @@ import type {
 } from '../../lib/client/page-context.js';
 import { type StaffEnrollment, StaffEnrollmentSchema } from '../../lib/client/safe-db-types.js';
 import { QueryClientProviderDebug } from '../../lib/client/tanstackQuery.js';
-import { getStudentDetailUrl } from '../../lib/client/url.js';
+import { getStudentEnrollmentUrl } from '../../lib/client/url.js';
 import type { EnumEnrollmentStatus } from '../../lib/db-types.js';
 
 import { ColumnManager } from './components/ColumnManager.js';
@@ -95,6 +95,7 @@ function StudentsCard({
 
   useEffect(() => {
     // TODO: This is a workaround to avoid a hydration mismatch.
+    // eslint-disable-next-line @eslint-react/hooks-extra/no-direct-set-state-in-use-effect
     setIsMediumOrLarger(mediaQuery?.matches ?? true);
   }, [mediaQuery]);
 
@@ -182,18 +183,20 @@ function StudentsCard({
       columnHelper.accessor((row) => row.user?.uid ?? row.enrollment.pending_uid, {
         id: 'user_uid',
         header: 'UID',
-        cell: (info) => info.getValue(),
+        cell: (info) => {
+          return (
+            <a href={getStudentEnrollmentUrl(urlPrefix, info.row.original.enrollment.id)}>
+              {info.getValue()}
+            </a>
+          );
+        },
       }),
       columnHelper.accessor((row) => row.user?.name, {
         id: 'user_name',
         header: 'Name',
         cell: (info) => {
           if (info.row.original.user) {
-            return (
-              <a href={getStudentDetailUrl(urlPrefix, info.row.original.user.user_id)}>
-                {info.getValue() || '—'}
-              </a>
-            );
+            return info.getValue() || '—';
           }
           return (
             <OverlayTrigger overlay={<Tooltip>Student information is not yet available.</Tooltip>}>
@@ -205,7 +208,7 @@ function StudentsCard({
       columnHelper.accessor((row) => row.enrollment.status, {
         id: 'enrollment_status',
         header: 'Status',
-        cell: (info) => <EnrollmentStatusIcon status={info.getValue()} />,
+        cell: (info) => <EnrollmentStatusIcon type="text" status={info.getValue()} />,
         filterFn: (row, columnId, filterValues: string[]) => {
           if (filterValues.length === 0) return true;
           const current = row.getValue(columnId);