+++ apps/prairielearn/src/tests/accessibility/index.test.ts
@@ -14,6 +14,7 @@ import { config } from '../../lib/config.js';
 import { features } from '../../lib/features/index.js';
 import { TEST_COURSE_PATH } from '../../lib/paths.js';
 import { assertNever } from '../../lib/types.js';
+import { ensureEnrollment } from '../../models/enrollment.js';
 import * as news_items from '../../news_items/index.js';
 import * as server from '../../server.js';
 import * as helperServer from '../helperServer.js';
@@ -28,7 +29,7 @@ const SITE_URL = 'http://localhost:' + config.serverPort;
 async function loadPageJsdom(url: string): Promise<{ text: string; jsdom: JSDOM }> {
   const text = await fetch(url).then((res) => {
     if (!res.ok) {
-      throw new Error(`Error loading page: ${res.status}`);
+      throw new Error(`Error loading page "${url}": ${res.status}`);
     }
     return res.text();
   });
@@ -421,6 +422,14 @@ describe('accessibility', () => {
       IdSchema,
     );
 
+    const enrollment = await ensureEnrollment({
+      course_instance_id: '1',
+      user_id,
+      agent_user_id: null,
+      agent_authn_user_id: null,
+      action_detail: 'implicit_joined',
+    });
+
     await features.enable('question-sharing');
 
     routeParams = {
@@ -429,6 +438,7 @@ describe('accessibility', () => {
       assessment_id,
       question_id,
       user_id,
+      enrollment_id: enrollment!.id,
     };
 
     await sqldb.executeRow('UPDATE questions SET share_publicly = true WHERE id = $question_id', {